% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prisma_dashboard.R
\name{Prisma_dashboard}
\alias{Prisma_dashboard}
\title{Harmonized PRiSMA Dashboard (Pregnancy Surveillance)}
\usage{
Prisma_dashboard(Prisma_dat, ...)
}
\arguments{
\item{Prisma_dat}{- Prisma dataset,}

\item{...}{- any additional function}
}
\description{
This function helps to create a shiny dashboard for you using Prisma data.
}
